Crate agent

Structs

InvalidMessage

Represents an error that occurs when decoding a message.

Metrics

Represents system metrics.

Reader

A struct for reading data from a byte slice.

Enums

Message

Represents a message that can be sent or received.

MessageType

Represents the type of message being sent or received.

Traits

Codec

Represents a type that can be encoded into and decoded from a byte stream.